scraper.py
import requests
from bs4 import BeautifulSoup
import trafilatura

class Scraper:
    def __init__(self):
        self.session = requests.Session()

    def get_metrics(self, platform, handle):
        url = self._get_profile_url(platform, handle)
        content = self._get_website_text_content(url)
        
        # Implement platform-specific scraping logic here
        # This is a placeholder and should be customized for each platform
        metrics = {
            'followers': self._extract_follower_count(content),
            'likes': self._extract_like_count(content),
            'comments': self._extract_comment_count(content)
        }
        
        return metrics

    def _get_profile_url(self, platform, handle):
        if platform == 'tiktok':
            return f"https://www.tiktok.com/@{handle}"
        elif platform == 'instagram':
            return f"https://www.instagram.com/{handle}/"
        # Add other platforms as needed
        else:
            raise ValueError(f"Unsupported platform for scraping: {platform}")

    def _get_website_text_content(self, url: str) -> str:
        downloaded = trafilatura.fetch_url(url)
        text = trafilatura.extract(downloaded)
        return text

    def _extract_follower_count(self, content):
        # Implement follower count extraction logic
        pass

    def _extract_like_count(self, content):
        # Implement like count extraction logic
        pass

    def _extract_comment_count(self, content):
        # Implement comment count extraction logic
        pass
