pyproject.toml
[tool.poetry]
name = "repl-nix-socialengageanalyzer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
trafilatura = "^1.12.2"
flask = "^3.0.3"
instagram-private-api = "^1.6.0.0"
tiktokapi = "^6.5.2"
flask-sqlalchemy = "^3.1.1"
flask-apscheduler = "^1.13.1"
tweepy = "^4.14.0"
facebook-sdk = "^3.1.0"
linkedin-api = "^2.3.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
